- moves java json serialization to its own library
