+++ serialization/java/json/lib/src/test/java/com/microsoft/kiota/serialization/json/LibraryTest.java
@@ -1,0 +1,13 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package com.microsoft.kiota.serialization.json;
+
+import org.junit.jupiter.api.Test;
+import static org.junit.jupiter.api.Assertions.*;
+
+class LibraryTest {
+    @Test void testSomeLibraryMethod() {
+        assertTrue(true, "someLibraryMethod should return 'true'");
+    }
+}
